+++ frontend/components/badge/styles.css
@@ -1,5 +1,4 @@
 .badge {
-  display: flex;
   align-items: center;
   justify-content: center;
   border: 1px solid black;
