+++ frontend/components/comments/styles.css
@@ -1,8 +1,6 @@
 .comments {
-  display: flex;
   flex-direction: column;
   position: relative;
   width: 940px;
-  height: 250px;
   border: 1px solid;
 }
