+++ frontend/components/company-contacts/component.jsx
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+import styles from './styles.css';
+
+export default class CompanyContacts extends Component {
+  render() {
+    return (
+      <div className={styles.contacts}>
+        CompanyContacts
+      </div>
+    );
+  }
+}
