+++ frontend/components/company-contacts/styles.css
@@ -1,4 +1,4 @@
-.location {
+.contacts {
   display: flex;
   flex-direction: column;
   position: relative;
