--- frontend/components/location-contacts/component.jsx
@@ -1,11 +1,0 @@
-import React, { Component } from 'react';
-import styles from './styles.css';
-
-export default class LocationContacts extends Component {
-  render() {
-    return (
-      <div className={styles.locationContacts} />
-    );
-  }
-}
-
