--- frontend/components/location-contacts/styles.css
@@ -1,5 +1,0 @@
-.locationContacts {
-  display: flex;
-  flex-grow: 1;
-  border: 1px solid;
-}
