--- frontend/components/location-map/component.jsx
@@ -1,11 +1,0 @@
-import React, { Component } from 'react';
-import styles from './styles.css';
-
-export default class LocationMap extends Component {
-  render() {
-    return (
-      <div className={styles.locationMap} />
-    );
-  }
-}
-
