--- frontend/components/location-map/styles.css
@@ -1,5 +1,0 @@
-.locationMap {
-  display: flex;
-  flex-grow: 2;
-  border: 1px solid;
-}
