--- frontend/components/quest-location/component.jsx
@@ -1,16 +1,0 @@
-import React, { Component } from 'react';
-import styles from './styles.css';
-
-export default class QuestLocation extends Component {
-  render() {
-    return (
-      <div className={styles.location}>
-        {this.props.children}
-      </div>
-    );
-  }
-}
-
-QuestLocation.propTypes = {
-  children: React.PropTypes.array.isRequired,
-};
