+++ frontend/components/similar/styles.css
@@ -3,6 +3,5 @@
   flex-direction: row;
   position: relative;
   width: 940px;
-  height: 250px;
   border: 1px solid;
 }
